FIxed bugs with memory allocation, norminette + extra files, redirection now working 95% of the times
