+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: fbarrett <fbarrett@student.42quebec>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 14:31:34 by fbarrett          #+#    #+#             */
-/*   Updated: 2024/01/16 15:46:17 by fbarrett         ###   ########.fr       */
+/*   Updated: 2024/01/17 14:29:24 by fbarrett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,165 +28,6 @@ int	seek_pipe(char	**line_args)
 	return (pipe_nbr);
 }
 
-int	here_doc(int file, char	*delimiter, int *a)
-{
-	char	*buff;
-	char	*here_doc_name;
-
-	here_doc_name = ft_calloc(3, sizeof(char));
-	here_doc_name[0] = '.';
-	here_doc_name[1] = 'A';
-	here_doc_name[2] = '\0';
-	if (!access(here_doc_name, F_OK))
-		here_doc_name[1] += 1;
-	file = open(here_doc_name, O_CREAT | O_WRONLY | O_TRUNC, 0000644);
-	if (file < 0)
-		return (file);
-	while (1)
-	{
-		buff = readline("Input something: ");
-		if (!buff)
-			return (-1);
-		if (!ft_strncmp(delimiter, buff, ft_strlen(delimiter)))
-			break ;
-		write(file, buff, ft_strlen(buff));
-		write(file, "\n", 1);
-		free(buff);
-	}
-	if (buff)
-		free(buff);
-	close(file);
-	file = open(here_doc_name, O_RDONLY, 0000644);
-	close(STDIN_FILENO);
-	dup2(file, STDIN_FILENO);
-	close(file);
-	a += 2;
-	free(here_doc_name);
-	return (file);
-}
-
-int	r_redirect(int file, char *given_file, int *a)
-{
-	file = open(given_file, O_CREAT | O_WRONLY | O_TRUNC, 0000644);
-	if (file < 0)
-		return (file);
-	close(STDOUT_FILENO);
-	dup2(file, STDOUT_FILENO);
-	close(file);
-	a += 2;
-   return (file);	
-}
-
-int	ra_redirect(int file, char *given_file, int *a)
-{
-	file = open(given_file, O_CREAT | O_WRONLY | O_APPEND, 0000644);
-	if (file < 0)
-		return (file);
-	close(STDOUT_FILENO);
-	dup2(file, STDOUT_FILENO);
-	close(file);
-	a += 2;
-   return (file);	
-}
-
-int	l_redirect(int file, char *given_file, int *a)
-{
-	file = open(given_file, O_RDONLY, 0000644);
-	if (file < 0)
-		return (file);
-	close(STDIN_FILENO);
-	dup2(file, STDIN_FILENO);
-	close(file);
-	a += 2;
-   return (file);	
-}
-
-
-int	check_redirection(char **line)
-{
-	int i;
-	int	file;
-	int a;
-
-	i = 0;
-	a = 0;
-	while (line[i])
-	{
-		if (!ft_strncmp(">", line[i], 2))
-		{
-			file = r_redirect(file, line[i + 1], &a);
-		}
-		else if (!ft_strncmp(">>", line[i], 3))
-		{
-			file = ra_redirect(file, line[i + 1], &a);
-		}
-		else if (!ft_strncmp("<", line[i], 2))
-		{
-			file = l_redirect(file, line[i + 1], &a);
-		}
-		else if (!ft_strncmp("<<", line[i], 3))
-		{
-			file = here_doc(file, line[i + 1], &a);
-		}
-		if (file < 0)
-		{
-			perror("Error");
-			return (-1);
-		}
-		i++;
-	}
-	return (i - a);
-}
-
-char	**line_args_parse(char **line, int args_nbr)
-{
-	char	**line_args;
-	int i;
-	int a;
-
-	i = 0;
-	a = 0;
-	line_args = ft_calloc(args_nbr + 1, sizeof(char *));
-	while (line[i])
-	{
-		if (!ft_strncmp("<", line[i], 2) || !ft_strncmp("<<", line[i], 3) || !ft_strncmp(">", line[i], 2) || !ft_strncmp(">>", line[i], 3))
-		{
-			i += 2;
-			a += 2;
-			continue ;
-		}
-		line_args[i - a] = ft_strdup(line[i]);
-		i++;
-	}
-	line_args[i + a] = 0;
-	return (line_args);
-}
-
-int	run_single_cmd(char	**line, char *cmd_path,	char **envp, s_pipe *pipe)
-{
-	char	**line_args;
-	int		line_args_nbr;
-	
-	if	((pipe->child = fork()) < 0)
-		return (1);
-	if (pipe->child > 0)
-	{
-		waitpid(pipe->child, &pipe->error, 0);
-		return (0);
-	}
-	if (!pipe->child)
-	{
-		line_args_nbr = check_redirection(line);
-		if (line_args_nbr < 0)
-			return (1);
-		line_args = line_args_parse(line, line_args_nbr);
-		if (execve(cmd_path, line_args, envp) == -1)
-			perror("execve failed to execute");
-		free(line_args);
-	}
-	exit (1);
-}
-
 int	run_each_cmd(s_pipe *pipes, char **cmd_paths, char **envp, char **line)
 {
 	char	**line_args;
@@ -204,10 +45,10 @@ int	run_each_cmd(s_pipe *pipes, char **cmd_paths, char **envp, char **line)
 		if (!pipes->child)
 		{
 			child_process(pipes, line);
-			line_args_nbr = check_redirection(line);
+			line_args_nbr = check_redirection(pipes->cmd_args);
 			if (line_args_nbr < 0)
 				return (1);
-			line_args = line_args_parse(pipes->cmd_args, line_args_nbr);
+			line_args = line_rm_redirection(pipes->cmd_args, line_args_nbr);
 			if (execve(cmd_paths[pipes->pipes_nbr - pipes->i], line_args, envp) == -1)
 				perror("execve failed to execute");	
 		}
@@ -217,11 +58,11 @@ int	run_each_cmd(s_pipe *pipes, char **cmd_paths, char **envp, char **line)
 	return (0);
 }
 
-int	run_cmds_pipe(char **line, char	**cmd_paths, char **envp, s_pipe *pipes)
+int	run_cmds(char **line, char	**cmd_paths, char **envp, s_pipe *pipes)
 {
 	pipes->i = 0;
 	pipes->cmd_ptr = 0;
-	pipes->child_list = ft_calloc(pipes->pipes_nbr + 2, sizeof(int));
+	pipes->child_list = ft_calloc(pipes->pipes_nbr + 2, sizeof(unsigned long int));
 	while (pipes->i < pipes->pipes_nbr)
 	{
 		pipe(pipes->fd);
@@ -264,40 +105,12 @@ int	exec_line(s_pipe *pipe, char **line_args, char **envp, char *buff)
 {
 	char	**cmd_paths;
 
-	pipe->pipes_nbr = seek_pipe(line_args);
-	if (pipe->pipes_nbr == 0)
-	{
-		cmd_paths = ft_calloc(2, sizeof(char *));
-		cmd_paths[0] = seek_cmd(line_args[0], envp);
-		cmd_paths[1] = 0;
-		if (cmd_paths[0])
-			run_single_cmd(line_args, cmd_paths[0], envp, pipe);
-	}
-	else
-	{
-		cmd_paths = ft_calloc(2 + pipe->pipes_nbr, sizeof(char *));
-		seek_all_cmds(&cmd_paths, line_args, envp);
-		if (!check_cmds(pipe, cmd_paths))
-			run_cmds_pipe(line_args, cmd_paths, envp, pipe);
-	}
+	cmd_paths = ft_calloc(2 + pipe->pipes_nbr, sizeof(unsigned long int));
+	seek_all_cmds(&cmd_paths, line_args, envp);
+	if (!check_cmds(pipe, cmd_paths))
+		run_cmds(line_args, cmd_paths, envp, pipe);
 	free_moi_ca(buff, cmd_paths, line_args);
 	return (0);
-}
-
-void	unlink_here_doc(void)
-{
-	char	*here_doc;
-
-	here_doc = ft_calloc(3, sizeof(char));
-	here_doc[0] = '.';
-	here_doc[1] = 'A';
-	here_doc[2] = '\0';
-	while (!access(here_doc, F_OK))
-	{
-		unlink(here_doc);
-		here_doc[1]++;
-	}
-	free(here_doc);
 }
 
 int main(int argc, char	**argv, char **envp)
@@ -308,7 +121,7 @@ int main(int argc, char	**argv, char **envp)
 
 	(void)argc;
 	(void)argv;
-	pipe = ft_calloc(1, sizeof(char *));
+	pipe = ft_calloc(1, sizeof(s_pipe));
 	while (1)
 	{
 		buff = readline("> ");
@@ -334,7 +147,6 @@ int main(int argc, char	**argv, char **envp)
 		unlink_here_doc();
 	}
 	free(pipe);
-	// y faut rl_clear_history
 	clear_history();
 	return (0);
 }
